+++ tests/Stub/Race.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Proxy\Tests\Stub;
+
+final class Race
+{
+    public function time(): int
+    {
+        return 10;
+    }
+}