+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.2 under development
 
-- no changes in this release.
+- Bug #58: Fix rendering of class modifiers (@arogachev)
 
 ## 1.0.1 July 11, 2022